apps/psychotherapy_tracker/pyproject.toml
[project]
name = "psychotherapy-tracker"
version = "0.1.0"
description = "Voice-based psychological wellness tracker with AI analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "plotly>=5.18.0",
    "azure-cognitiveservices-speech>=1.34.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv]
dev-dependencies = []
